# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "向前端提供api",
        "title": "online practice server",
        "contact": {
            "name": "软工小组",
            "url": "http://www.swagger.io/support",
            "email": "2020111058@email.szu.edu.cn"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
        "/login": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "Login generates the authentication token",
                "parameters": [
                    {
                        "description": "id or email",
                        "name": "username",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Password",
                        "name": "password",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":errno,\"message\":\"err_msg\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/users/register": {
            "post": {
                "description": "输入账号密码，并选择角色",
                "produces": [
                    "application/json"
                ],
                "summary": "注册",
                "parameters": [
                    {
                        "description": "{",
                        "name": "regRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.RegRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":errno,\"message\":\"err_msg\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/users/test": {
            "get": {
                "tags": [
                    "测试"
                ],
                "summary": "测试各个模块间信息能否流通",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "user.RegRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "role": {
                    "type": "integer"
                },
                "schoolId": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    }
}